perf: use spatial sort on point indices
Constructing the Delaunay triangulation was previously slowed down
because the index data had to be added to the vertices of the
triangulation after construction. The index was obtained by first
sorting and searching through the points by coordinate. This commit adds
indexing data to the vertices during construction of the triangulation
by inserting the points one at a time and immediately updating the
vertex handles, but spatial sorts the points before insertion to
preserve the speed benefits of group insertion.